move 'react-ga-proxy.js' to 'common' dir
