+++ source/js/buyers-guide/components/social-share/social-share.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import ReactGA from "../../../react-ga-proxy";
+import ReactGA from "../../../common/react-ga-proxy";
 import copyToClipboard from "../../../../js/copy-to-clipboard.js";
 
 const SocialShareLink = props => {
