+++ source/js/common/bind-event-handlers.js
@@ -1,4 +1,4 @@
-import ReactGA from "../react-ga-proxy.js";
+import ReactGA from "./react-ga-proxy.js";
 
 /**
  * Bind click handler to #donate-footer-btn
