+++ source/js/components/join/join.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import ReactGA from "../../react-ga-proxy.js";
+import ReactGA from "../../common/react-ga-proxy.js";
 import ReactDOM from "react-dom";
 import classNames from "classnames";
 import CountrySelect from "../petition/country-select.jsx";
